jpms: Refactor handlebars.jackson2
