+++ handlebars-jackson2/src/test/java/com/github/jknack/handlebars/jackson2/Comment.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jknack.handlebars;
+package com.github.jknack.handlebars.jackson2;
 
 public class Comment {
   private String author;
